Address PR review feedback
- Use Where().OrderByAsArray() for non-destructive filtering in GetDiagnostics
- Revert uint back to int for consistency with Roslyn
- Fix copypasta bug in WithSuppressUniqueIds (RootNamespace -> SuppressUniqueIds)
- Remove trailing space in debugger display string
- Extract CreateDescriptor helper and refactor all descriptors in RazorDiagnostics.cs
- Make lambdas static in IncrementalValueProviderExtensions
- Use PooledArrayBuilder(capacity: 2) in source generator RazorProviders
- Extract ParseRazorLanguageVersion and ParseRazorWarningLevel helper methods
- Distinguish absent vs empty RazorWarningLevel: absent (old SDK) defaults to 0,
  empty (new SDK) defaults to language version
- Fix encoding issues in comments (em dash -> dash)

+++ src/Compiler/Microsoft.CodeAnalysis.Razor.Compiler/src/Language/RazorLanguageVersion.cs
@@ -118,5 +118,5 @@ public sealed record RazorLanguageVersion : IComparable<RazorLanguageVersion>
     /// The warning level corresponds to the major version number
     /// (e.g., <see cref="Version_11_0"/> → <c>11</c>).
     /// </summary>
-    public uint GetDefaultWarningLevel() => (uint)Major;
+    public int GetDefaultWarningLevel() => Major;
 }
